Revert "import, epxort work"
This reverts commit 34d849752514ff848dad7e4201b180bd10c7fd64.

+++ FinanceAppGUI/ITransactionLogic.cs
@@ -7,6 +7,6 @@
         void RemoveIncome(Transaction t);
         void RemoveExpense(Transaction t);
         void SetupCollections(IList<Transaction> expense, IList<Transaction> income);
-        double Balance();
+        int Balance();
     }
 }
